Add hide_relativenumbers option
Turns off relativenumber in the current window while peeking.

The relativenumber option is useful for jumping around a file with `5j`
or `7k` and requires less math in one's head to use. Hoever if one is
using `:<number>` they are switching to a mode where they only care
about absolute line numbers, so it would be useful for the number line
to also switch to absolute line numbers.

A new option `hide_relativenumbers` option is added so that users who
don't like this behavior can easily turn it off. However, I left it on
by default since I think it provides the best user experience to the
most users.